Update patch set 1
Patch Set 1:

Linking to previous discussion on this:
http://review.cyanogenmod.org/#/c/121998/

The original upstream commit where make_f2fs was added was here:
https://android.googlesource.com/platform/external/f2fs-tools/+/3686eecd8c8511ae7ecd7611a2f065b48b484b13%5E%21/

I guess as long as recovery still contains a usable mkfs.f2fs, then this gerrit is okay.  Although the only remaining mkfs.f2fs module is now a host executable, which is perhaps a bit unintuitive (for someone looking at this after the fact).

Patch-set: 1