Update patch set 1
Patch Set 1:

Are you still interested in these Daniel? If so I'm about to bring up a rebased tree that compiles and brings it to 1.6.0. If you bring the kernel up to 3.4.y on f2fs-stable the f2fs-tools to 1.6.0 + dev branch it may work.

However, I'm not sure said tree will be ready to push to CM yet. I don't believe it should break anything for older f2fs versions but I haven't tested yet.

Patch-set: 1