Update patch set 3
Patch Set 3:

I think the purpose of the code in fsck is to detect possible changes in the f2fs driver to optimize/fix partitions. I'm honestly not sure of this, I haven't looked into this very much, but this seemed a reasonable explanation as to why one would check the kernel version.

Now, I've seen many backport f2fs from recent kernel releases, which means that f2fs gets heavily modified without the kernel version being changed. If the local version was not appended to the kernel version and we skip the timestamp, fsck would never be able to detect possible changes to the f2fs driver.

This is why I decided to make this opt-in.

Patch-set: 3